Added simple test script
This test server can be started with 'node example/test.js'.

Currently it is very simple, it just prints out the process id back
to the requester. One the console it logs the PID, URL and the
connection's socket's file descriptor.

Once a baseline of the server is in place, this will use the file
descriptor passed in by node-launcher, rather than creating the
socket.