Conditionalise spawn too fast support
The spawning too fast support doesn't work very well in practise.
In production mode it fails to respawn when it should; in debug mode
respawn on error has proved to be not a very useful model.

For now it has been conditionalised, but may be removed completely in
the future.